Refactor Steppable and offender sar form related parts based on feedback
